refactor: move directive helper
