+++ example/src/main/kotlin/com.expedia.graphql.sample/extension/CustomSchemaGeneratorHooks.kt
@@ -1,6 +1,6 @@
 package com.expedia.graphql.sample.extension
 
-import com.expedia.graphql.generator.directive.DirectiveWiringHelper
+import com.expedia.graphql.DirectiveWiringHelper
 import com.expedia.graphql.hooks.DataFetcherExecutionPredicate
 import com.expedia.graphql.hooks.SchemaGeneratorHooks
 import com.expedia.graphql.sample.validation.DataFetcherExecutionValidator
